requirements.txt
alembic==1.0.0
click==6.7
Flask==1.1.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10.1
Mako==1.0.7
MarkupSafe==1.1.1
psycopg2-binary==2.9.3
python-dateutil==2.7.3
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.11
Werkzeug==1.0.1
pyjwt==2.3.0
pandas==1.3.5
